changed IQuery from an interface to a class
